// packages/frontend/src/api/client/models/DocumentDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ScraperShortDto } from './ScraperShortDto';
export type DocumentDto = {
    id: string;
    title: string;
    content: string;
    scraper: ScraperShortDto;
    scrapedAt: string;
    contentHash: string;
};